factory: Allow informers to be stopped
By passing in a stop channel we can (should we choose) stop the
informers running. This is useful in unit tests where you want to run:

 go test -race -count 10000 -test.run TestEndpointSlice

but that fails with:

 race: limit on 8128 simultaneously alive goroutines is exceeded, dying

which makes it hard to ensure you've actually fixed a race because
quite often this is triggered after, say, 100 invocations of the test.