Make HandleEndpointSlice call through HandleEndpoints
The aggregated EndpointSlice is effectively the same as you would get
in a call to HandleEndpoints(). In unit tests we want to assert that
plugin.HandleEndpoints() is called so delegate to HandleEndpoints().
The alternative is to extend the plugin interface to accommodate
endpointslices but that is a change that will ripple out to all the
other plugins and is not needed at the moment.