Ensure idled services work with endpointslices
An `oc idle <service>` will annotate the endpoint with an 'idle time'
annotation. We don't or can't add this annotation to endpointslices in
openshift/oc (idle) because the underlying endpointslices are managed
by the kube-controller-manager and that doesn't propagate annotations
as part of the mirroring service.

We explicitly check to see if the underlying endpoint for the service
has the idle annotation when processing endpointslices. There's
probably a cost here which mitigates the switch to endpointslices
somewhat. With that in place, any time a service is idled the
endpointslices will also carry the annotation and this allows us to
correctly write the "ept:" value to haproxy.config.

At the moment this seems to be the cheapest change to ensure that
idled services are honoured when we're listening to endpointslices.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1900989